修改密码、个人信息、退出登入功能添加

+++ plan-admin/src/main/java/com/xinghui/ctrl/mgt/AuthCtrl.java
@@ -1,10 +1,15 @@
 package com.xinghui.ctrl.mgt;
 
+import com.xinghui.config.GlobalException;
+import com.xinghui.constants.ResultCode;
 import com.xinghui.dto.LoginInfoDTO;
 import com.xinghui.dto.RegisterDTO;
+import com.xinghui.dto.UpdatePassWordDTO;
+import com.xinghui.entity.AccountInfoDO;
 import com.xinghui.entity.SysUserDO;
 import com.xinghui.enums.TerminalTypeEnum;
 import com.xinghui.service.AccountInfoService;
+import com.xinghui.service.SysUserService;
 import com.xinghui.utils.RequestContextUtil;
 import com.xinghui.utils.ResponseUtil;
 import com.xinghui.utils.ResultDTO;
@@ -13,6 +18,8 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
 import ma.glasnost.orika.MapperFacade;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -31,6 +38,12 @@ public class AuthCtrl {
 
     @Resource
     private MapperFacade mapperFacade;
+
+    @Resource
+    private SysUserService sysUserService;
+
+    @Resource
+    private BCryptPasswordEncoder encoder;
 
     @PostMapping("/v1/login")
     @ApiOperation(value = "登入")
@@ -57,6 +70,31 @@ public class AuthCtrl {
     @PostMapping("/v1/logout")
     @ApiOperation(value = "撤销")
     public ResultDTO logout() {
+        accountService.removeToken(RequestContextUtil.userId(), RequestContextUtil.terminal());
+        return ResponseUtil.success(true);
+    }
+
+    @PostMapping("/v1/update-password")
+    @ApiOperation(value = "修改密码")
+    public ResultDTO updatePassWord(@RequestBody UpdatePassWordDTO updatePassWordDTO) {
+        if (StringUtils.isBlank(updatePassWordDTO.getNewPassword())) {
+            throw new GlobalException(ResultCode.NEW_PASSWORD_NOT_NULL);
+        }
+        if (StringUtils.isBlank(updatePassWordDTO.getConfirmNewPassword())) {
+            throw new GlobalException(ResultCode.CONFIRM_NEW_PASSWORD_NOT_NULL);
+        }
+        if (!updatePassWordDTO.getNewPassword().equals(updatePassWordDTO.getConfirmNewPassword())) {
+            throw new GlobalException(ResultCode.NOT_CONFIRM_NEW_PASSWORD);
+        }
+
+        //获取人员信息
+        SysUserDO sysUserDO = sysUserService.getById(RequestContextUtil.userId());
+        //获取登入信息
+        AccountInfoDO accountInfoDO = accountService.getById(sysUserDO.getAccountId());
+        accountInfoDO.setPassword(encoder.encode(updatePassWordDTO.getConfirmNewPassword()));
+        //更新密码
+        accountService.updateById(accountInfoDO);
+        //撤销token
         accountService.removeToken(RequestContextUtil.userId(), RequestContextUtil.terminal());
         return ResponseUtil.success(true);
     }
